+++ run.py
@@ -186,12 +186,6 @@ def get_depot_tools(source_dir, fetch=False):
     return dir
 
 
-PATCH_INFO = {
-    "macos_screen_capture.patch": (2, []),
-    "macos_use_xcode_clang.patch": (1, ["build"]),
-    "windows_fix_optional.patch": (1, ["third_party"]),
-}
-
 PATCHES = {
     "windows_x86_64": [
         "4k.patch",
@@ -340,6 +334,31 @@ def apply_patch(patch, dir, depth):
                 cmd(["patch", f"-p{depth}"], stdin=stdin)
 
 
+def _deps_dirs(src_dir):
+    cap = cmdcap(["gclient", "recurse", "-j1", "pwd"])
+    abs_dirs = cap.split("\n")
+    rel_dirs = [os.path.relpath(abs_dir, src_dir) for abs_dir in abs_dirs]
+    return rel_dirs
+
+
+def apply_patches(target, patch_dir, src_dir):
+    with cd(src_dir):
+        for patch in PATCHES[target]:
+            apply_patch(os.path.join(patch_dir, patch), src_dir, 1)
+            cmd(["gclient", "recurse", "git", "add", "--", ":!*.orig"])
+            cmd(
+                [
+                    "gclient",
+                    "recurse",
+                    "git",
+                    "commit",
+                    "--allow-empty",
+                    "-am",
+                    f"[shiguredo-patch] Apply {patch}",
+                ]
+            )
+
+
 def get_webrtc(source_dir, patch_dir, version, target, webrtc_source_dir=None):
     if webrtc_source_dir is None:
         webrtc_source_dir = os.path.join(source_dir, "webrtc")
@@ -363,10 +382,7 @@ def get_webrtc(source_dir, patch_dir, version, target, webrtc_source_dir=None):
             cmd(["git", "checkout", "-f", version])
             cmd(["git", "clean", "-df"])
             cmd(["gclient", "sync", "-D", "--force", "--reset", "--with_branch_heads"])
-            for patch in PATCHES[target]:
-                depth, dirs = PATCH_INFO.get(patch, (1, ["."]))
-                dir = os.path.join(src_dir, *dirs)
-                apply_patch(os.path.join(patch_dir, patch), dir, depth)
+            apply_patches(target, patch_dir, src_dir)
 
 
 def fetch_webrtc(source_dir, patch_dir, version, target, webrtc_source_dir=None):
@@ -379,10 +395,7 @@ def fetch_webrtc(source_dir, patch_dir, version, target, webrtc_source_dir=None)
         cmd(["git", "checkout", "-f", version])
         cmd(["git", "clean", "-df"])
         cmd(["gclient", "sync", "-D", "--force", "--reset", "--with_branch_heads"])
-        for patch in PATCHES[target]:
-            depth, dirs = PATCH_INFO.get(patch, (1, ["."]))
-            dir = os.path.join(src_dir, *dirs)
-            apply_patch(os.path.join(patch_dir, patch), dir, depth)
+        apply_patches(target, patch_dir, src_dir)
 
 
 def revert_webrtc(source_dir, patch_dir, target, webrtc_source_dir=None):
@@ -391,12 +404,44 @@ def revert_webrtc(source_dir, patch_dir, target, webrtc_source_dir=None):
 
     src_dir = os.path.join(webrtc_source_dir, "src")
     with cd(src_dir):
+        dirs = _deps_dirs(src_dir)
+        for dir in dirs:
+            with cd(dir):
+                # 時雨堂パッチが当たっていない最新のコミットを取得する
+                lines = cmdcap(["git", "log", "--oneline", "-n30"]).split("\n")
+                for line in lines:
+                    if "[shiguredo-patch]" in line:
+                        continue
+                    commit = line.split(" ")[0]
+                    break
+                cmd(["git", "reset", "--soft", commit])
         cmd(["gclient", "recurse", "git", "reset", "--hard"])
         cmd(["gclient", "recurse", "git", "clean", "-df"])
-        for patch in PATCHES[target]:
-            depth, dirs = PATCH_INFO.get(patch, (1, ["."]))
-            dir = os.path.join(src_dir, *dirs)
-            apply_patch(os.path.join(patch_dir, patch), dir, depth)
+        apply_patches(target, patch_dir, src_dir)
+
+
+def diff_webrtc(source_dir, webrtc_source_dir=None):
+    if webrtc_source_dir is None:
+        webrtc_source_dir = os.path.join(source_dir, "webrtc")
+
+    src_dir = os.path.join(webrtc_source_dir, "src")
+    with cd(src_dir):
+        cmd(["gclient", "recurse", "git", "add", "-N", "--", ":!*.orig"])
+        dirs = _deps_dirs(src_dir)
+        for dir in dirs:
+            with cd(dir):
+                cmd(
+                    [
+                        "git",
+                        "--no-pager",
+                        "diff",
+                        "--ignore-submodules",
+                        "--src-prefix",
+                        f"a/{dir}/".replace("\\", "/"),
+                        "--dst-prefix",
+                        f"b/{dir}/".replace("\\", "/"),
+                    ]
+                )
 
 
 def git_get_url_and_revision(dir):
@@ -1368,6 +1413,15 @@ def main():
     rp.add_argument("--build-dir")
     rp.add_argument("--webrtc-source-dir")
     rp.add_argument("--webrtc-build-dir")
+    # ソースコードの差分を出力する
+    dp = sp.add_parser("diff")
+    dp.set_defaults(op="diff")
+    dp.add_argument("target", choices=TARGETS)
+    dp.add_argument("--debug", action="store_true")
+    dp.add_argument("--source-dir")
+    dp.add_argument("--build-dir")
+    dp.add_argument("--webrtc-source-dir")
+    dp.add_argument("--webrtc-build-dir")
     # 現在 build と package を分ける意味は無いのだけど、
     # 今後複数のビルドを纏めてパッケージングする時に備えて別コマンドにしておく
     pp = sp.add_parser("package")
@@ -1546,6 +1600,17 @@ def main():
                 source_dir,
                 patch_dir,
                 args.target,
+                webrtc_source_dir=webrtc_source_dir,
+            )
+
+    if args.op == "diff":
+        mkdir_p(source_dir)
+
+        with cd(BASE_DIR):
+            dir = get_depot_tools(source_dir, fetch=False)
+            add_path(dir)
+            diff_webrtc(
+                source_dir,
                 webrtc_source_dir=webrtc_source_dir,
             )